removed useless object from sqFileChooser and moved ShowFileSave into EditorView
