+++ src/main/scala/view/util/scalaQuestSwingComponents/SqSwingFileChooser.scala
@@ -7,21 +7,6 @@ import javax.swing.filechooser.FileNameExtensionFilter
 import javax.swing.{JComponent, JFileChooser}
 
 /**
- * Utility object used to get and show elements in a FileChooser.
- */
-object SqSwingFileChooser {
-
-  def showFileSave(title: String, onSave: String => Unit, selectedFileName: String, parent: JComponent): Unit = {
-    val chooser = SqSwingFileChooser(title)
-    chooser.setSelectedFile(new File(selectedFileName))
-    if (chooser.showSaveDialog(parent) == JFileChooser.APPROVE_OPTION) {
-      onSave(chooser.getSelectedFile.getPath)
-    }
-  }
-
-}
-
-/**
  * Represents a custom FileChooser for ScalaQuest.
  *
  * @param title the title of the FileChooser.
@@ -35,4 +20,7 @@ case class SqSwingFileChooser(title: String) extends JFileChooser {
       ResourceName.FileExtensions.StoryFileExtension
     )
   )
+
+
+
 }
